Move copied delta-rs methods to a separate module
