+++ src/context.rs
@@ -16,7 +16,7 @@ use tokio::io::{AsyncReadExt, AsyncWriteExt, BufReader};
 use std::fs::File;
 
 use datafusion::datasource::listing::{
-    ListingOptions, ListingTable, ListingTableConfig, ListingTableUrl, PartitionedFile,
+    ListingOptions, ListingTable, ListingTableConfig, ListingTableUrl,
 };
 use datafusion::datasource::object_store::ObjectStoreUrl;
 use datafusion::execution::context::{default_session_builder, SessionState};
@@ -34,17 +34,16 @@ use futures::{StreamExt, TryStreamExt};
 
 #[cfg(test)]
 use mockall::automock;
-use object_store::{path::Path, ObjectMeta, ObjectStore};
+use object_store::{path::Path, ObjectStore};
 
 use sqlparser::ast::{
     AlterTableOperation, CreateFunctionBody, FunctionDefinition, Ident, ObjectName,
     ObjectType, SchemaName, Statement, TableFactor, TableWithJoins,
 };
 
-use arrow::compute::{cast_with_options, CastOptions};
 use arrow_integration_test::field_to_json;
-use arrow_schema::{ArrowError, DataType, TimeUnit};
-use chrono::{DateTime, FixedOffset, NaiveDateTime, Utc};
+use arrow_schema::DataType;
+use chrono::{DateTime, FixedOffset, Utc};
 use std::iter::zip;
 use std::ops::Deref;
 use std::str::FromStr;
@@ -64,7 +63,6 @@ use datafusion::optimizer::{OptimizerContext, OptimizerRule};
 use datafusion::physical_expr::create_physical_expr;
 use datafusion::physical_expr::execution_props::ExecutionProps;
 use datafusion::physical_optimizer::pruning::PruningPredicate;
-use datafusion::physical_plan::file_format::{partition_type_wrap, FileScanConfig};
 use datafusion::physical_plan::filter::FilterExec;
 use datafusion::physical_plan::projection::ProjectionExec;
 use datafusion::prelude::SessionConfig;
@@ -91,24 +89,20 @@ use datafusion_expr::logical_plan::{
     CreateCatalog, CreateCatalogSchema, CreateExternalTable, CreateMemoryTable,
     DropTable, Extension, LogicalPlan, Projection,
 };
-use datafusion_expr::{DmlStatement, Expr, Filter, WriteOp};
+use datafusion_expr::{DmlStatement, Filter, WriteOp};
 use deltalake::action::{Action, Add, Remove};
-use deltalake::delta_datafusion::DeltaDataChecker;
-use deltalake::operations::writer::{DeltaWriter, WriterConfig};
 use deltalake::operations::{create::CreateBuilder, write::WriteBuilder};
-use deltalake::storage::DeltaObjectStore;
-use deltalake::{DeltaResult, DeltaTable, Schema as DeltaSchema};
+use deltalake::{DeltaTable, Schema as DeltaSchema};
 use log::{debug, info, warn};
-use object_store::path::DELIMITER;
 use parking_lot::RwLock;
 use prost::Message;
 use tempfile::TempPath;
 use tokio::sync::Semaphore;
-use url::Url;
 use uuid::Uuid;
 
 use crate::catalog::{PartitionCatalog, DEFAULT_SCHEMA, STAGING_SCHEMA};
 use crate::datafusion::visit::VisitorMut;
+use crate::delta_rs::backports::{parquet_scan_from_actions, write_execution_plan};
 #[cfg(test)]
 use crate::frontend::http::tests::deterministic_uuid;
 use crate::provider::{
@@ -530,231 +524,6 @@ pub fn is_statement_read_only(statement: &DFStatement) -> bool {
     }
 }
 
-// Appropriated from https://github.com/delta-io/delta-rs/pull/1176; once the DELETE and UPDATE ops
-// are available through delta-rs this will be obsolete.
-/// Write the provide ExecutionPlan to the underlying storage
-/// The table's invariants are checked during this proccess
-pub async fn write_execution_plan(
-    table: &DeltaTable,
-    state: SessionState,
-    plan: Arc<dyn ExecutionPlan>,
-    partition_columns: Vec<String>,
-    object_store: Arc<DeltaObjectStore>,
-    target_file_size: Option<usize>,
-    write_batch_size: Option<usize>,
-) -> Result<Vec<Add>> {
-    let invariants = table
-        .get_metadata()
-        .and_then(|meta| meta.schema.get_invariants())
-        .unwrap_or_default();
-    let checker = DeltaDataChecker::new(invariants);
-
-    // Write data to disk
-    let mut tasks = vec![];
-    for i in 0..plan.output_partitioning().partition_count() {
-        let inner_plan = plan.clone();
-        let task_ctx = Arc::new(TaskContext::from(&state));
-
-        let config = WriterConfig::new(
-            inner_plan.schema(),
-            partition_columns.clone(),
-            None,
-            target_file_size,
-            write_batch_size,
-        );
-        let mut writer = DeltaWriter::new(object_store.clone(), config);
-        let checker_stream = checker.clone();
-        let mut stream = inner_plan.execute(i, task_ctx)?;
-        let handle: tokio::task::JoinHandle<DeltaResult<Vec<Add>>> =
-            tokio::task::spawn(async move {
-                while let Some(maybe_batch) = stream.next().await {
-                    let batch = maybe_batch?;
-                    checker_stream.check_batch(&batch).await?;
-                    writer.write(&batch).await?;
-                }
-                writer.close().await
-            });
-
-        tasks.push(handle);
-    }
-
-    // Collect add actions to add to commit
-    Ok(futures::future::join_all(tasks)
-        .await
-        .into_iter()
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(|err| {
-            DataFusionError::Execution(format!(
-                "Failed writing to delta table {table}: {err}"
-            ))
-        })?
-        .into_iter()
-        .collect::<Result<Vec<_>, _>>()?
-        .concat()
-        .into_iter()
-        .collect::<Vec<_>>())
-}
-
-// Appropriated from https://github.com/delta-io/delta-rs/pull/1176 with minor changes.
-// Once the DELETE and UPDATE ops are available through delta-rs this will be obsolete.
-/// Create a Parquet scan limited to a set of files
-pub async fn parquet_scan_from_actions(
-    table: &DeltaTable,
-    actions: &[Add],
-    schema: &Schema,
-    filters: &[Expr],
-    state: &SessionState,
-    projection: Option<&Vec<usize>>,
-    limit: Option<usize>,
-) -> Result<Arc<dyn ExecutionPlan>> {
-    // TODO we group files together by their partition values. If the table is partitioned
-    // and partitions are somewhat evenly distributed, probably not the worst choice ...
-    // However we may want to do some additional balancing in case we are far off from the above.
-    let mut file_groups: HashMap<Vec<ScalarValue>, Vec<PartitionedFile>> = HashMap::new();
-    for action in actions {
-        let part = partitioned_file_from_action(action, schema);
-        file_groups
-            .entry(part.partition_values.clone())
-            .or_default()
-            .push(part);
-    }
-
-    let table_partition_cols = table.get_metadata()?.partition_columns.clone();
-    let file_schema = Arc::new(Schema::new(
-        schema
-            .fields()
-            .iter()
-            .filter(|f| !table_partition_cols.contains(f.name()))
-            .cloned()
-            .collect(),
-    ));
-
-    let url = Url::parse(&table.table_uri()).unwrap();
-    let host = format!(
-        "{}-{}{}",
-        url.scheme(),
-        url.host_str().unwrap_or_default(),
-        url.path().replace(DELIMITER, "-").replace(':', "-")
-    );
-    state
-        .runtime_env()
-        .register_object_store("delta-rs", &host, table.object_store());
-    let object_store_url = ObjectStoreUrl::parse(format!("delta-rs://{host}"))?;
-
-    ParquetFormat::new()
-        .create_physical_plan(
-            state,
-            FileScanConfig {
-                object_store_url,
-                file_schema,
-                file_groups: file_groups.into_values().collect(),
-                statistics: table.datafusion_table_statistics(),
-                projection: projection.cloned(),
-                limit,
-                table_partition_cols: table_partition_cols
-                    .iter()
-                    .map(|c| {
-                        Ok((
-                            c.to_owned(),
-                            partition_type_wrap(
-                                schema.field_with_name(c)?.data_type().clone(),
-                            ),
-                        ))
-                    })
-                    .collect::<Result<Vec<_>, ArrowError>>()?,
-                output_ordering: None,
-                infinite_source: false,
-            },
-            filters,
-        )
-        .await
-}
-
-// Copied from delta-rs as it's private there; once the DELETE and UPDATE ops
-// are available through delta-rs this will be obsolete.
-fn partitioned_file_from_action(action: &Add, schema: &Schema) -> PartitionedFile {
-    let partition_values = schema
-        .fields()
-        .iter()
-        .filter_map(|f| {
-            action.partition_values.get(f.name()).map(|val| match val {
-                Some(value) => to_correct_scalar_value(
-                    &serde_json::Value::String(value.to_string()),
-                    f.data_type(),
-                )
-                .unwrap_or(ScalarValue::Null),
-                None => ScalarValue::Null,
-            })
-        })
-        .collect::<Vec<_>>();
-
-    let ts_secs = action.modification_time / 1000;
-    let ts_ns = (action.modification_time % 1000) * 1_000_000;
-    let last_modified = DateTime::<Utc>::from_utc(
-        NaiveDateTime::from_timestamp_opt(ts_secs, ts_ns as u32).unwrap(),
-        Utc,
-    );
-    PartitionedFile {
-        object_meta: ObjectMeta {
-            location: Path::from(action.path.clone()),
-            last_modified,
-            size: action.size as usize,
-        },
-        partition_values,
-        range: None,
-        extensions: None,
-    }
-}
-
-// Copied from delta-rs as it's private there; once the DELETE and UPDATE ops
-// are available through delta-rs this will be obsolete.
-fn to_correct_scalar_value(
-    stat_val: &serde_json::Value,
-    field_dt: &DataType,
-) -> Option<ScalarValue> {
-    match stat_val {
-        serde_json::Value::Array(_) => None,
-        serde_json::Value::Object(_) => None,
-        serde_json::Value::Null => None,
-        serde_json::Value::String(string_val) => match field_dt {
-            DataType::Timestamp(_, _) => {
-                let time_nanos = ScalarValue::try_from_string(
-                    string_val.to_owned(),
-                    &DataType::Timestamp(TimeUnit::Nanosecond, None),
-                )
-                .ok()?;
-                let cast_arr = cast_with_options(
-                    &time_nanos.to_array(),
-                    field_dt,
-                    &CastOptions { safe: false },
-                )
-                .ok()?;
-                Some(ScalarValue::try_from_array(&cast_arr, 0).ok()?)
-            }
-            _ => {
-                Some(ScalarValue::try_from_string(string_val.to_owned(), field_dt).ok()?)
-            }
-        },
-        other => match field_dt {
-            DataType::Timestamp(_, _) => {
-                let time_nanos = ScalarValue::try_from_string(
-                    other.to_string(),
-                    &DataType::Timestamp(TimeUnit::Nanosecond, None),
-                )
-                .ok()?;
-                let cast_arr = cast_with_options(
-                    &time_nanos.to_array(),
-                    field_dt,
-                    &CastOptions { safe: false },
-                )
-                .ok()?;
-                Some(ScalarValue::try_from_array(&cast_arr, 0).ok()?)
-            }
-            _ => Some(ScalarValue::try_from_string(other.to_string(), field_dt).ok()?),
-        },
-    }
-}
-
 // The only reason to keep this trait around (instead of migrating all the functions directly into
 // DefaultSeafowlContext), is that `create_physical_plan` would then be a recursive async function,
 // which works for traits, but not for structs: https://stackoverflow.com/a/74737853
@@ -1088,7 +857,7 @@ impl DefaultSeafowlContext {
             .await?;
 
         // TODO: if `DeltaTable::get_version_timestamp` was globally public we could also pass the
-        // exact version timestamp, instead of creating onw automatically in our own catalog (which
+        // exact version timestamp, instead of creating one automatically in our own catalog (which
         // could lead to minor timestamp differences).
         self.table_catalog
             .create_new_table_version(table_uuid, table.version())
@@ -1675,7 +1444,7 @@ impl SeafowlContext for DefaultSeafowlContext {
                 // This is actually CREATE TABLE AS
                 let physical = self.create_physical_plan(input).await?;
 
-                // First create the table and then insert the data from the subqeury
+                // First create the table and then insert the data from the subquery
                 // TODO: this means we'll have 2 table versions at the end, 1st from the create
                 // and 2nd from the insert, while it seems more reasonable that in this case we have
                 // only one
@@ -2070,10 +1839,6 @@ impl SeafowlContext for DefaultSeafowlContext {
                                 .get_collection_id_by_name(&self.database, name)
                                 .await?
                             {
-                                // TODO: as for DROP TABLE, we should really only flag all tables
-                                // in the schema as DROPED, and then try to lazy delete them during
-                                // subsequent `VACUUM`s. Only once all the underlying table files
-                                // are deleted is it safe to drop the collection.
                                 self.table_catalog.drop_collection(collection_id).await?
                             };
