+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod config;
 pub mod context;
 pub mod data_types;
 pub mod datafusion;
+pub mod delta_rs;
 pub mod frontend;
 pub mod nodes;
 pub mod object_store;
